General bug fixes after the merge
