+++ src/main/java/SFWE405/DataSeeder.java
@@ -1,4 +1,4 @@
-package SFWE405.Project;
+package SFWE405;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
