+++ src/main/java/SFWE405/Project/repository/HomeworkAssignmentsRepository.java
@@ -1,7 +1,7 @@
 package SFWE405.Project.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import SFWE405.Project.entity.HomeworkAssignments;
+import SFWE405.Project.entity.HomeworkAssignment;
 
-public interface HomeworkAssignmentsRepository extends JpaRepository<HomeworkAssignments, Long> {
+public interface HomeworkAssignmentsRepository extends JpaRepository<HomeworkAssignment, Long> {
 }
