+++ src/main/java/SFWE405/Project/repository/SemestersRepository.java
@@ -1,8 +1,8 @@
 package SFWE405.Project.repository;
-import SFWE405.Project.entity.Semesters;
+import SFWE405.Project.entity.Semester;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SemestersRepository extends JpaRepository<Semesters, Long> {
+public interface SemestersRepository extends JpaRepository<Semester, Long> {
 }
