+++ src/main/java/SFWE405/ProjectApplication.java
@@ -1,4 +1,4 @@
-package SFWE405.Project;
+package SFWE405;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
